requirements.frozen.txt
# Web Framework and Extensions
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Werkzeug==2.0.3
WTForms==3.2.1

# Database
SQLAlchemy==1.4.23
alembic==1.14.0
psycopg2-binary==2.9.9
greenlet==3.1.1

# AI and API Integration
openai>=1.0.0
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
tenacity==8.2.3

# Text Processing and NLP
newspaper3k==0.2.8
nltk==3.8.1
fuzzywuzzy==0.18.0
Levenshtein==0.25.1
python-Levenshtein==0.25.1
jieba3k==0.35.1
tinysegmenter==0.3
regex==2024.4.16

# Async and Network
aiohttp==3.10.5
aiohappyeyeballs==2.4.0
aiosignal==1.3.1
eventlet==0.33.3
python-socketio==5.5.0
python-engineio==4.3.0

# Task Queue and Message Broker
celery==5.2.3
redis==4.1.0
amqp==5.2.0
kombu==5.4.2
billiard==3.6.4.0
vine==5.1.0

# Data Processing
pandas==2.2.2
numpy==1.24.3
openpyxl==3.1.2
pydantic==2.9.2
pydantic_core==2.23.4

# HTML/XML Processing
beautifulsoup4==4.12.3
lxml==5.2.1
lxml_html_clean==0.1.1
cssselect==1.2.0

# Utils
python-dotenv==1.0.1
tqdm==4.66.2
Pillow==10.4.0
svgwrite==1.4.3
Tree==0.2.4
PyYAML==6.0.2

# Date and Time
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.2

# Dependencies and Others
attrs==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.4.1
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
rapidfuzz==3.8.1
requests==2.31.0
requests-file==2.1.0
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.12.1